@@ -4,54 +4,47 @@
 [project]
 name = "damply"
 version = "0.9.0"
-description = "A Quick Tool For Sorting Dicom Files"
+description = "Data Management Plan made simple."
 license = "MIT"
 readme = "README.md"
 keywords = ["damply", "pixi", "python", "package"]
-
 authors = [{ name = "Jermiah Joseph", email = "jermiahjoseph98@gmail.com" }]
 maintainers = [{ name = "Jermiah Joseph", email = "jermiahjoseph98@gmail.com" }]
-
-requires-python = ">= 3.10"
+requires-python = ">= 3.11"
 dependencies = [
-  "rich",
-  "rich-click",
-  "pydantic",
-  "textual",
-  "trogon",
-  "dataclasses-json",
-  "click-didyoumean",
-  "matplotlib",
-  "plotly",
-  "pandas", "kaleido", "pygetent",
+    "rich",
+    "rich-click",
+    "pydantic",
+    "dataclasses-json",
+    "matplotlib",
+    "plotly",
+    "pandas",
+    "pygetent",
+    "pybytesize>=0.8.1,<0.9",
 ]
 classifiers = [
-  "License :: OSI Approved :: MIT License",
-  "Natural Language :: English",
-  "Operating System :: OS Independent",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-  "Programming Language :: Python :: 3.12",
+    "License :: OSI Approved :: MIT License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 
 [project.urls]
-homepage = "https://github.com/jjjermiah/damply"
-repository = "https://github.com/jjjermiah/damply"
-documentation = "https://jjjermiah.github.io/damply/"
-changelog = "https://github.com/jjjermiah/damply/blob/main/docs/CHANGELOG.md"
-issues = "https://github.com/jjjermiah/damply/issues"
+homepage = "https://github.com/bhklab/damply"
+repository = "https://github.com/bhklab/damply"
+documentation = "https://bhklab.github.io/damply/"
+changelog = "https://github.com/bhklab/damply/blob/main/docs/CHANGELOG.md"
+issues = "https://github.com/bhklab/damply/issues"
 
 [project.scripts]
 dmp = "damply.cli.main:cli"
 damplyplot = "damply.plot:main"
 
-[project.optional-dependencies]
-style = ["rich-codex"]
-
 #--------------------------------------------------------------------------------------------------#
 ############################################## PIXI ################################################
 #__________________________________________________________________________________________________#
-
 [tool.pixi.system-requirements]
 linux = "3.10.0"
 
@@ -69,19 +62,19 @@ damply = { path = ".", editable = true }
 [tool.pixi.environments]
 dev = { features = ["test", "style", "docs"], solve-group = "default" }
 publish = { features = ["build", "release"], solve-group = "default" }
-py310 = ["py310", "test"]
 py311 = ["py311", "test"]
 py312 = ["py312", "test"]
+py313 = ["py313", "test"]
 
 ############################################## python ###############################################
-
-[tool.pixi.feature.py310.dependencies]
-python = "3.10.*"
 [tool.pixi.feature.py311.dependencies]
 python = "3.11.*"
+
 [tool.pixi.feature.py312.dependencies]
 python = "3.12.*"
 
+[tool.pixi.feature.py313.dependencies]
+python = "3.13.*"
 ############################################## TEST ################################################
 [tool.pixi.feature.test.dependencies]
 pytest = "*"
