@@ -1,4 +1,3 @@
 from .directory_list import Directory, DirectoryList
-from .byte_size import ByteSize
 
-__all__ = ["Directory", "DirectoryList", "ByteSize"]
+__all__ = ['Directory', 'DirectoryList']
