@@ -1,35 +1,33 @@
-import os
-from pathlib import Path
-
-
-import os
-from pathlib import Path
 import platform
+from pathlib import Path
+from typing import NoReturn
 
-def get_file_owner_full_name(file_path: Path):
-    try:
-        if platform.system() == "Windows":
-            raise NotImplementedError("Platform not supported for retrieving user info using pwd module.")
-        
-        from pwd import getpwuid
 
-        # Get the file's status
-        file_stat = os.stat(file_path)
+def _raise_windows_not_supported() -> NoReturn:
+	msg = 'Platform not supported for retrieving user info using pwd module.'
+	raise NotImplementedError(msg)
 
-        # Get the user ID of the file owner
-        uid = file_stat.st_uid
 
-        # Get the user information based on the user ID
-        user_info = getpwuid(uid)
+def get_file_owner_full_name(file_path: Path) -> str:
+	try:
+		if platform.system() == 'Windows':
+			_raise_windows_not_supported()
 
-        # Return the full name of the user
-        return user_info.pw_gecos
+		from pwd import getpwuid
 
-    except ImportError:
-        print("Module 'pwd' is not available on this platform.")
-        return file_path.owner()
-    except NotImplementedError:
-        return "Retrieving user info is not supported on Windows."
-    except Exception as e:
-        return str(e)
+		# Get the file's status
+		file_stat = file_path.stat()
+		# Get the user ID of the file owner
+		uid: int = file_stat.st_uid
 
+		# Get the user information based on the user ID
+		getpwuid(uid)
+
+	except ImportError:
+		return file_path.owner()
+	except NotImplementedError:
+		return 'Retrieving user info is not supported on Windows.'
+	except Exception as e:
+		return str(e)
+	else:
+		return getpwuid(uid).pw_gecos
